+++ NEWS.md
@@ -2,9 +2,13 @@
 
 * `fgts_dygraph()` ignores `splitcols` argument if only one series is plotted.
 * Cache created with .onLoad error
-* Fixed crash when the date column is part of the plot formula
 * New function `fg_get_datemap()` gets current date map used.
 * New function `fg_current_theme()` gets current `ggplot2` theme used.
+
+## Bug fixes
+
+* Fixed crash when the date column is part of the plot formula
+* Fixed crash when `event_ds` is an empty data.table
 
 # FinanceGraphs 0.9.0